move common methods to common/ua_utils.py
